fix: no longer depends on fastify or express types
I was able to implement a generic type for `NestCookieRequest`
that can be extended to work with `Request` and `FastfyRequest`
without worrying about having both installed for the module to work
in the first place. This will allow for greater freedom to devs
so they don't have to bloat their dependencies with unnecessary
packages.

fix: #2