ci: add semantic-release workflow
`semantic-release` will allow for the automatic publishing of
the package via any push to `main` that has a commit with the
`fix:`, `feat:`, or `BREAKING CHANGE` keywords in the commit
subject or body. This means the workflow will change a little
bit, ensuring that work is saved to a staging branch before being
pushed all together to `main`, but that shouldn't be a big deal.
Dependabot should not be affected by this.